backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-decouple>=3.8
dj-database-url>=2.1
whitenoise>=6.6
gunicorn>=21.2
